Move ds9 meta translation
